% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_auto.R
\name{attach_spdf}
\alias{attach_spdf}
\title{Attach an "sf" object to a grid}
\usage{
attach_spdf(x, spdf)
}
\arguments{
\item{x}{object to attach "sf" object to}

\item{spdf}{an "sf" object to attach}
}
\description{
Attach an "sf" object to a grid
}
