% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{aus_grid1}
\alias{aus_grid1}
\title{aus_grid1}
\usage{
aus_grid1
}
\description{
A data frame containing a grid layout for the Australian States and Territories, thanks to \url{https://github.com/jonocarroll}.
}
\keyword{data}
