% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{aus_pop}
\alias{aus_pop}
\title{aus_pop}
\usage{
aus_pop
}
\description{
March 2017 population data for Australian states and territories by age group. Online source no longer available.
}
\keyword{data}
