% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{auto_states}
\alias{auto_states}
\alias{auto_countries}
\title{auto_states}
\description{
List of valid values for countries for fetching rnaturalearth data when used with \code{\link{grid_auto}} to create a grid of states.

List of valid values for continents for fetching rnaturalearth data when used with \code{\link{grid_auto}} to create a grid of countries.
}
\keyword{data}
