% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_grid.R
\name{create_new_grid}
\alias{create_new_grid}
\title{Create an order-preserving, roughly square grid based on lat/long}
\usage{
create_new_grid(x, y, unit.name, num.iterations = 100)
}
\arguments{
\item{x}{a vector of unit longitudes}

\item{y}{a vector of unit latitude}

\item{unit.name}{a vector of names}

\item{num.iterations}{an integer of how many times to try shrinking}
}
\description{
Create an order-preserving, roughly square grid based on lat/long
}
