% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{election}
\alias{election}
\title{election}
\usage{
election
}
\description{
2016 US presidential election results, obtained from \url{https://docs.google.com/spreadsheets/d/133Eb4qQmOxNvtesw2hdVns073R68EZx4SfCnP4IGQf8/htmlview?sle=true}.
}
\keyword{data}
