% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_geo.R
\name{facet_geo}
\alias{facet_geo}
\title{Arrange a sequence of geographical panels into a grid that preserves some geographical orientation}
\usage{
facet_geo(facets, ..., grid = "us_state_grid1", move_axes = TRUE)
}
\arguments{
\item{facets}{passed to \code{\link[ggplot2]{facet_wrap}}}

\item{grid}{character vector of the grid layout to use (currently only "us_state_grid1" and "us_state_grid2" are available)}

\item{move_axes}{should axis labels and ticks be moved to the closest panel along the margins?}

\item{\ldots}{additional parameters passed to \code{\link[ggplot2]{facet_wrap}}}
}
\description{
Arrange a sequence of geographical panels into a grid that preserves some geographical orientation
}
\examples{
library(ggplot2)

# barchart of state rankings in various categories
ggplot(state_ranks, aes(variable, rank, fill = variable)) +
  geom_col() +
  coord_flip() +
  facet_geo(~ state) +
  theme_bw()

# use an alternative US state grid and place
ggplot(state_ranks, aes(variable, rank, fill = variable)) +
  geom_col() +
  coord_flip() +
  facet_geo(~ state, grid = "us_state_grid2") +
  theme(panel.spacing = unit(0.1, "lines"))

# use a free x-axis (not a good idea but just to show it works)
ggplot(state_ranks, aes(variable, rank, fill = variable)) +
  geom_col() +
  coord_flip() +
  facet_geo(~ state, scales = "free_x") +
  theme_bw()

# plot unemployment rate time series for each state
ggplot(state_unemp, aes(year, rate)) +
  geom_line() +
  facet_geo(~ state) +
  scale_x_continuous(labels = function(x) paste0("'", substr(x, 3, 4))) +
  ylab("Unemployment Rate (\%)")

# plot the 2016 unemployment rate
ggplot(subset(state_unemp, year == 2016), aes(factor(year), rate)) +
  geom_col(fill = "steelblue") +
  facet_geo(~ state) +
  theme(
    axis.title.x = element_blank(),
    axis.text.x = element_blank(),
    axis.ticks.x = element_blank()) +
  ylab("Unemployment Rate (\%)")
}
