% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geofacet-package.R
\docType{package}
\name{geofacet-package}
\alias{geofacet}
\alias{geofacet-package}
\title{geofacet: 'ggplot2' Faceting Utilities for Geographical Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides geographical faceting functionality for 'ggplot2'. Geographical faceting arranges a sequence of plots of data for different geographical entities into a grid that preserves some of the geographical orientation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hafen/geofacet}
  \item \url{https://hafen.github.io/geofacet/}
  \item Report bugs at \url{https://github.com/hafen/geofacet/issues}
}

}
\author{
\strong{Maintainer}: Ryan Hafen \email{rhafen@gmail.com}

Other contributors:
\itemize{
  \item Barret Schloerke \email{schloerke@gmail.com} [contributor]
}

}
\keyword{internal}
