% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_geo.R
\name{get_geofacet_grob}
\alias{get_geofacet_grob}
\title{Perform post-processing on a facet_geo ggplot object}
\usage{
get_geofacet_grob(x)
}
\arguments{
\item{x}{object of class 'facet_geo'}
}
\description{
Perform post-processing on a facet_geo ggplot object
}
