% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_auto.R
\name{get_ne_data}
\alias{get_ne_data}
\title{Get rnaturalearth data}
\usage{
get_ne_data(code)
}
\arguments{
\item{code}{A country/continent name to get rnaturalearth data from (see \code{\link{auto_countries}} or \code{\link{auto_states}}).}
}
\description{
Get rnaturalearth data
}
\examples{
\dontrun{
dat <- get_ne_data("brazil")
}
}
