% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_geo.R
\name{grid_preview}
\alias{grid_preview}
\title{Plot a preview of a grid}
\usage{
grid_preview(x, use_code = TRUE)
}
\arguments{
\item{x}{a data frame containing a grid}

\item{use_code}{use the code variable to determine text labels? If FALSE, the geographical entity name is used instead.}
}
\description{
Plot a preview of a grid
}
\examples{
grid_preview(us_state_grid2)
grid_preview(eu_grid1, use_code = FALSE)
}
