% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_docs.R
\docType{data}
\name{grids}
\alias{grids}
\alias{us_state_grid1}
\alias{us_state_grid2}
\alias{eu_grid1}
\alias{aus_grid1}
\alias{sa_prov_grid1}
\alias{gb_london_boroughs_grid}
\alias{nhs_scot_grid}
\alias{india_grid1}
\alias{india_grid2}
\alias{argentina_grid1}
\alias{br_states_grid1}
\alias{sea_grid1}
\alias{mys_grid1}
\alias{fr_regions_grid1}
\alias{de_states_grid1}
\alias{us_or_counties_grid1}
\alias{us_wa_counties_grid1}
\alias{us_in_counties_grid1}
\alias{us_in_central_counties_grid1}
\alias{se_counties_grid1}
\alias{sf_bay_area_counties_grid1}
\alias{ua_region_grid1}
\alias{mx_state_grid1}
\alias{mx_state_grid2}
\alias{scotland_local_authority_grid1}
\alias{us_state_grid3}
\alias{italy_grid1}
\alias{italy_grid2}
\alias{be_province_grid1}
\alias{us_state_grid4}
\alias{jp_prefs_grid1}
\alias{ng_state_grid1}
\alias{bd_upazila_grid1}
\alias{spain_prov_grid1}
\alias{ch_cantons_grid1}
\alias{ch_cantons_grid2}
\alias{china_prov_grid1}
\alias{world_86countries_grid}
\alias{se_counties_grid2}
\alias{uk_regions1}
\alias{us_state_contiguous_grid1}
\alias{sk_province_grid1}
\alias{ch_aargau_districts_grid1}
\alias{jo_gov_grid1}
\alias{spain_ccaa_grid1}
\alias{spain_prov_grid2}
\alias{world_countries_grid1}
\alias{br_states_grid2}
\alias{china_city_grid1}
\alias{kr_seoul_district_grid1}
\alias{nz_regions_grid1}
\alias{sl_regions_grid1}
\alias{us_census_div_grid1}
\alias{ar_tucuman_province_grid1}
\alias{us_nh_counties_grid1}
\alias{china_prov_grid2}
\alias{pl_voivodeships_grid1}
\alias{us_ia_counties_grid1}
\alias{us_id_counties_grid1}
\alias{ar_cordoba_dep_grid1}
\alias{us_fl_counties_grid1}
\alias{ar_buenosaires_communes_grid1}
\alias{nz_regions_grid2}
\alias{oecd_grid1}
\alias{ec_prov_grid1}
\alias{nl_prov_grid1}
\alias{ca_prov_grid1}
\alias{us_nc_counties_grid1}
\alias{mx_ciudad_prov_grid1}
\alias{bg_prov_grid1}
\alias{us_hhs_regions_grid1}
\alias{tw_counties_grid1}
\alias{tw_counties_grid2}
\alias{af_prov_grid1}
\alias{us_mi_counties_grid1}
\alias{pe_prov_grid1}
\alias{sa_prov_grid2}
\alias{mx_state_grid3}
\alias{cn_bj_districts_grid1}
\alias{us_va_counties_grid1}
\alias{us_mo_counties_grid1}
\alias{cl_santiago_prov_grid1}
\alias{us_tx_capcog_counties_grid1}
\alias{sg_planning_area_grid1}
\alias{in_state_ut_grid1}
\alias{cn_fujian_prov_grid1}
\alias{ca_quebec_electoral_districts_grid1}
\alias{nl_prov_grid2}
\alias{cn_bj_districts_grid2}
\alias{ar_santiago_del_estero_prov_grid1}
\alias{ar_formosa_prov_grid1}
\alias{ar_chaco_prov_grid1}
\alias{ar_catamarca_prov_grid1}
\alias{ar_jujuy_prov_grid1}
\alias{ar_neuquen_prov_grid1}
\alias{ar_san_luis_prov_grid1}
\alias{ar_san_juan_prov_grid1}
\alias{ar_santa_fe_prov_grid1}
\alias{ar_la_rioja_prov_grid1}
\alias{ar_mendoza_prov_grid1}
\alias{ar_salta_prov_grid1}
\alias{ar_rio_negro_prov_grid1}
\alias{uy_departamentos_grid1}
\alias{ar_buenos_aires_prov_electoral_dist_grid1}
\alias{europe_countries_grid1}
\alias{argentina_grid2}
\alias{us_state_grid5}
\alias{jp_prefs_grid2}
\title{Geo Grids}
\usage{
us_state_grid1

us_state_grid2

eu_grid1

aus_grid1

sa_prov_grid1

gb_london_boroughs_grid

nhs_scot_grid

india_grid1

india_grid2

argentina_grid1

br_states_grid1

sea_grid1

mys_grid1

fr_regions_grid1

de_states_grid1

us_or_counties_grid1

us_wa_counties_grid1

us_in_counties_grid1

us_in_central_counties_grid1

se_counties_grid1

sf_bay_area_counties_grid1

ua_region_grid1

mx_state_grid1

mx_state_grid2

scotland_local_authority_grid1

us_state_grid3

italy_grid1

italy_grid2

be_province_grid1

us_state_grid4

jp_prefs_grid1

ng_state_grid1

bd_upazila_grid1

spain_prov_grid1

ch_cantons_grid1

ch_cantons_grid2

china_prov_grid1

world_86countries_grid

se_counties_grid2

uk_regions1

us_state_contiguous_grid1

sk_province_grid1

ch_aargau_districts_grid1

jo_gov_grid1

spain_ccaa_grid1

spain_prov_grid2

world_countries_grid1

br_states_grid2

china_city_grid1

kr_seoul_district_grid1

nz_regions_grid1

sl_regions_grid1

us_census_div_grid1

ar_tucuman_province_grid1

us_nh_counties_grid1

china_prov_grid2

pl_voivodeships_grid1

us_ia_counties_grid1

us_id_counties_grid1

ar_cordoba_dep_grid1

us_fl_counties_grid1

ar_buenosaires_communes_grid1

nz_regions_grid2

oecd_grid1

ec_prov_grid1

nl_prov_grid1

ca_prov_grid1

us_nc_counties_grid1

mx_ciudad_prov_grid1

bg_prov_grid1

us_hhs_regions_grid1

tw_counties_grid1

tw_counties_grid2

af_prov_grid1

us_mi_counties_grid1

pe_prov_grid1

sa_prov_grid2

mx_state_grid3

cn_bj_districts_grid1

us_va_counties_grid1

us_mo_counties_grid1

cl_santiago_prov_grid1

us_tx_capcog_counties_grid1

sg_planning_area_grid1

in_state_ut_grid1

cn_fujian_prov_grid1

ca_quebec_electoral_districts_grid1

nl_prov_grid2

cn_bj_districts_grid2

ar_santiago_del_estero_prov_grid1

ar_formosa_prov_grid1

ar_chaco_prov_grid1

ar_catamarca_prov_grid1

ar_jujuy_prov_grid1

ar_neuquen_prov_grid1

ar_san_luis_prov_grid1

ar_san_juan_prov_grid1

ar_santa_fe_prov_grid1

ar_la_rioja_prov_grid1

ar_mendoza_prov_grid1

ar_salta_prov_grid1

ar_rio_negro_prov_grid1

uy_departamentos_grid1

ar_buenos_aires_prov_electoral_dist_grid1

europe_countries_grid1

argentina_grid2

us_state_grid5

jp_prefs_grid2
}
\description{
There are now 108 grids available in this package and more online. To view a full list of available grids, see \href{https://raw.githubusercontent.com/hafen/grid-designer/master/grid_list.json}{here}. To create and submit your own grid, see \href{https://hafen.github.io/grid-designer/}{here}. To see several examples of grids being used to visualize data, see \code{\link{facet_geo}}.

\itemize{
\item \strong{us_state_grid1:} Grid layout for US states (including DC) Image reference \href{https://images7.alphacoders.com/687/687665.jpg}{here}.
}

\itemize{
\item \strong{us_state_grid2:} Grid layout for US states (including DC) Image reference \href{https://images7.alphacoders.com/687/687665.jpg}{here}.
}

\itemize{
\item \strong{eu_grid1:} Grid layout for the 28 EU Countries Image reference \href{https://ichef.bbci.co.uk/news/660/media/images/70233000/gif/_70233868_eunames.gif}{here}.
}

\itemize{
\item \strong{aus_grid1:} Grid layout for the Australian States and Territories. Image reference \href{https://cloud.githubusercontent.com/assets/9496865/26334627/75a729d2-3fa5-11e7-88af-a1dd86b7ddf3.png}{here}. Thanks to \href{https://github.com/jonocarroll}{jonocarroll}.
}

\itemize{
\item \strong{sa_prov_grid1:} Grid layout for the provinces of South Africa Image reference \href{https://camo.githubusercontent.com/14e357c36b6b72c0ef60e700d832b57aac87bddd/68747470733a2f2f7777772e736f7574686166726963612e746f2f70726f76696e6365732f5253415f62795f70726f76696e6365732e6a7067}{here}. Thanks to \href{https://github.com/jonmcalder}{jonmcalder}.
}

\itemize{
\item \strong{gb_london_boroughs_grid:} Grid layout for the boroughs of London. Note that the column \code{code_ons} contains the \href{http://geoportal.statistics.gov.uk/datasets/464be6191a434a91a5fa2f52c7433333_0?uiTab=table}{codes} used by UK Office for National Statistics. Image reference \href{https://cloud.githubusercontent.com/assets/13021448/26399784/3133d684-4076-11e7-8eac-1e74d023bf18.jpg}{here}. Thanks to \href{https://github.com/eldenvo}{eldenvo}.
}

\itemize{
\item \strong{nhs_scot_grid:} Grid layout for a grid of NHS Scotland Health Boards. Note that the column \code{code} contains the \href{http://geoportal.statistics.gov.uk/datasets/61dca36aa33e4aee99bbfa784fa9eb25_0?uiTab=table}{codes} used by UK Office for National Statistics. Image reference \href{https://cloud.githubusercontent.com/assets/24879214/26426113/253658d0-40cf-11e7-88b5-fe5ee7fda44f.png}{here}. Thanks to \href{https://github.com/jsphdms}{jsphdms}.
}

\itemize{
\item \strong{india_grid1:} Grid layout for India states (not including union territories). Image reference \href{https://cloud.githubusercontent.com/assets/11698124/26569168/dd2ca86c-44bd-11e7-9d74-af919522f7b1.jpg}{here}. Thanks to \href{https://github.com/meysubb}{meysubb}.
}

\itemize{
\item \strong{india_grid2:} Grid layout for India states (not including union territories). Image reference \href{https://cloud.githubusercontent.com/assets/11698124/26569168/dd2ca86c-44bd-11e7-9d74-af919522f7b1.jpg}{here}.
}

\itemize{
\item \strong{argentina_grid1:} Grid for the 23 provinces of Argentina. It includes the Malvinas/Falkland Islands and the Antarctic Territories (these are disputed, but they are included since many researchers might use data from these locations). Image reference \href{https://s-media-cache-ak0.pinimg.com/736x/bf/cb/96/bfcb9654477c127d3b0e0b72d5be65fc.jpg}{here}. Thanks to \href{https://github.com/eliocamp}{eliocamp}.
}

\itemize{
\item \strong{br_states_grid1:} Grid for the 27 states of Brazil. Image reference \href{https://camo.githubusercontent.com/477627dce2610f18f1bf3850cd32fca01610656e/687474703a2f2f63656a656475696e66616e74696c2e7062776f726b732e636f6d2f662f313435383034313938342f4d61706142726173696c5369676c61732e676966}{here}. Thanks to \href{https://github.com/italocegatta}{italocegatta}.
}

\itemize{
\item \strong{sea_grid1:} Grid for South East Asian countries. Image reference \href{https://camo.githubusercontent.com/694a275db0e89984f347ce8b49221352416d2c48/68747470733a2f2f7669676e65747465322e77696b69612e6e6f636f6f6b69652e6e65742f757461752f696d616765732f382f38322f536f757468656173745f617369612e6a70672f7265766973696f6e2f6c61746573743f63623d3230313130313230303431383034}{here}. Thanks to \href{https://github.com/jasonjb82}{jasonjb82}.
}

\itemize{
\item \strong{mys_grid1:} Grid for Malaysian states and territories. Image reference \href{https://camo.githubusercontent.com/cd22848f0b7b53d4ef7afc6d178a0e9579ecf2aa/68747470733a2f2f7777772e6d6170736f66776f726c642e636f6d2f6d616c61797369612f6d6170732f6d616c61797369612d706f6c69746963616c2d6d61702e6a7067}{here}. Thanks to \href{https://github.com/jasonjb82}{jasonjb82}.
}

\itemize{
\item \strong{fr_regions_grid1:} Land and overseas regions of France. Codes are INSEE codes. Image reference \href{https://camo.githubusercontent.com/b526c357a4c30cdd2e90700c6559381348111377/687474703a2f2f7777772e6361727465736672616e63652e66722f6361727465732f67656f677261706869652f63617274652d726567696f6e732e6a7067}{here}. Thanks to \href{https://github.com/mtmx}{mtmx}.
}

\itemize{
\item \strong{de_states_grid1:} Grid for the German states ('Länder') Image reference \href{https://camo.githubusercontent.com/e144ebb2bdd2d82d8a4c34908cd68ab2b2388425/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f642f64332f5374617465735f6f665f4765726d616e792e737667}{here}. Thanks to \href{https://github.com/DominikVogel}{DominikVogel}.
}

\itemize{
\item \strong{us_or_counties_grid1:} Grid for Oregon counties. Image reference \href{https://user-images.githubusercontent.com/11638312/27445795-01b77804-5730-11e7-919b-36cde8027706.jpg}{here}. Thanks to \href{https://github.com/aosmith16}{aosmith16}.
}

\itemize{
\item \strong{us_wa_counties_grid1:} Grid for Washington counties. Image reference \href{http://www.countymapsofwashington.com/aapics/washingstate.gif}{here}.
}

\itemize{
\item \strong{us_in_counties_grid1:} Grid for Indiana counties. Image reference \href{https://upload.wikimedia.org/wikipedia/commons/thumb/e/ec/Indiana_county_map.png/640px-Indiana_county_map.png}{here}. Thanks to \href{https://github.com/nateapathy}{nateapathy}.
}

\itemize{
\item \strong{us_in_central_counties_grid1:} Grid for central Indiana counties. Image reference \href{https://camo.githubusercontent.com/eb31e821890a0c600f59234fb184335c2bf84423/687474703a2f2f7777772e736176692e6f72672f77702d636f6e74656e742f75706c6f6164732f323031352f30322f6d61702d312e6a7067}{here}. Thanks to \href{https://github.com/nateapathy}{nateapathy}.
}

\itemize{
\item \strong{se_counties_grid1:} Grid for counties of Sweden. Image reference \href{http://www.worldatlas.com/webimage/countrys/europe/lgcolor/secounties.gif}{here}. Thanks to \href{https://github.com/duleise}{duleise}.
}

\itemize{
\item \strong{sf_bay_area_counties_grid1:} Grid of the 9 San Francisco Bay Area counties. Image reference \href{https://user-images.githubusercontent.com/1275592/28043128-53be650a-6585-11e7-920e-4ad9a13bf1a5.png}{here}. Thanks to \href{https://github.com/Eunoia}{Eunoia}.
}

\itemize{
\item \strong{ua_region_grid1:} Grid of administrative divisions of Ukraine (24 oblasts, one autonomous region, and two cities). Image reference \href{https://camo.githubusercontent.com/0615706478de63da2c7c3c3b21fe92c6db6a2fee/687474703a2f2f696d61676573312e66616e706f702e636f6d2f696d616765732f696d6167655f75706c6f6164732f556b7261696e69616e2d6d61702d756b7261696e652d313135333138375f3733305f3530352e676966}{here}. Thanks to \href{https://github.com/woldemarg}{woldemarg}.
}

\itemize{
\item \strong{mx_state_grid1:} Grid layout for the states of Mexico. Image reference \href{https://user-images.githubusercontent.com/1275592/28044555-ee236b2a-658c-11e7-96e5-3af124ca9619.png}{here}. Thanks to \href{https://github.com/ikashnitsky}{ikashnitsky}.
}

\itemize{
\item \strong{mx_state_grid2:} Grid layout for the states of Mexico. Image reference \href{https://user-images.githubusercontent.com/1275592/28044555-ee236b2a-658c-11e7-96e5-3af124ca9619.png}{here}. Thanks to \href{https://github.com/diegovalle}{diegovalle}.
}

\itemize{
\item \strong{scotland_local_authority_grid1:} Grid layout for the local authorities of Scotland. Image reference \href{https://user-images.githubusercontent.com/1275592/28044967-3b508570-658f-11e7-99b5-b1d73c765c2e.png}{here}. Thanks to \href{https://github.com/davidhen}{davidhen}.
}

\itemize{
\item \strong{us_state_grid3:} Grid layout for US states (excluding DC) Image reference \href{https://images7.alphacoders.com/687/687665.jpg}{here}. Thanks to \href{https://github.com/ejr248}{ejr248}.
}

\itemize{
\item \strong{italy_grid1:} Grid layout for regions of Italy (in collaboration with Stella Cangelosi and Luciana Dalla Valle). Image reference \href{https://user-images.githubusercontent.com/1275592/28291015-8f0abd66-6afd-11e7-9e2a-b4c1d4021c13.png}{here}. Thanks to \href{https://github.com/JulianStander}{JulianStander}.
}

\itemize{
\item \strong{italy_grid2:} Grid layout for regions of Italy (in collaboration with Stella Cangelosi and Luciana Dalla Valle). Image reference \href{https://user-images.githubusercontent.com/1275592/28291015-8f0abd66-6afd-11e7-9e2a-b4c1d4021c13.png}{here}. Thanks to \href{https://github.com/JulianStander}{JulianStander}.
}

\itemize{
\item \strong{be_province_grid1:} Grid layout for provinces of Belgium plus Brussels, including names in three languages (French, Dutch, English) and Belgium internal codes (NIS). Image reference \href{https://camo.githubusercontent.com/efa2c894793585141da8f7b9789b184eacc8d781/687474703a2f2f7777772e657570656469612e636f6d2f696d616765732f636f6e74656e742f42656c6769756d5f526567696f6e732e676966}{here}. Thanks to \href{https://github.com/ericlecoutre}{ericlecoutre}.
}

\itemize{
\item \strong{us_state_grid4:} Grid layout for US states (including DC). Image reference \href{https://images7.alphacoders.com/687/687665.jpg}{here}. Thanks to \href{https://github.com/kanishkamisra}{kanishkamisra}.
}

\itemize{
\item \strong{jp_prefs_grid1:} Grid layout for the prefectures of Japan. Image reference \href{https://camo.githubusercontent.com/bffec5ac4ae6507c1ca1b83570544f405b86b048/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f382f38372f49534f2d333136362d322d4a502e706e672f35313270782d49534f2d333136362d322d4a502e706e67}{here}. Thanks to \href{https://github.com/uribo}{uribo}.
}

\itemize{
\item \strong{ng_state_grid1:} Grid layout for the 37 Federal States of Nigeria. Image reference \href{https://camo.githubusercontent.com/fa12bca8fe3f3de3d7af707400adbc7fc4656cc3/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f382f38332f4e6967657269615f7374617465732e706e67}{here}. Thanks to \href{https://github.com/ghosthedirewolf}{ghosthedirewolf}.
}

\itemize{
\item \strong{bd_upazila_grid1:} Grid layout for Bangladesh 64 Upazilas. Image reference \href{https://camo.githubusercontent.com/df1e4aff5a3e0238afef97586d324d9b23acf27b/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f322f32382f42445f4d61705f61646d696e2e7376672f38303070782d42445f4d61705f61646d696e2e7376672e706e67}{here}. Thanks to \href{https://github.com/ghosthedirewolf}{ghosthedirewolf}.
}

\itemize{
\item \strong{spain_prov_grid1:} Grid layout for Provinces of Spain. Image reference \href{https://user-images.githubusercontent.com/1275592/30189589-cbfbf2de-93e9-11e7-9598-bb4c4e09e0c1.png}{here}. Thanks to \href{https://github.com/kintero}{kintero}.
}

\itemize{
\item \strong{ch_cantons_grid1:} Grid layout for Cantons of Switzerland. Image reference \href{https://camo.githubusercontent.com/b3c99e5725cdb60fd60a66731c8492fdb0e09ab6/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f342f34612f537769747a65726c616e642532435f61646d696e6973747261746976655f6469766973696f6e735f2d5f64655f2d5f636f6c6f7265642e737667}{here}. Thanks to \href{https://github.com/tinu-schneider}{tinu-schneider}.
}

\itemize{
\item \strong{ch_cantons_grid2:} Grid layout for Cantons of Switzerland. Image reference \href{https://user-images.githubusercontent.com/4510815/29034578-0485df5c-7b99-11e7-9ff8-67a7b687de12.png}{here}. Thanks to \href{https://github.com/rastrau}{rastrau}.
}

\itemize{
\item \strong{china_prov_grid1:} Grid layout for Provinces of China. Image reference \href{https://user-images.githubusercontent.com/1275592/30039720-cf9c8532-9189-11e7-856b-a6ecf449ee0c.png}{here}. Thanks to \href{https://github.com/weiyunna}{weiyunna}.
}

\itemize{
\item \strong{world_86countries_grid:} Grid layout for 86 countries in the world. Image reference \href{https://user-images.githubusercontent.com/1275592/30190118-0678aeb8-93ed-11e7-91b2-49738beb6f04.png}{here}. Thanks to \href{https://github.com/akangsha}{akangsha}.
}

\itemize{
\item \strong{se_counties_grid2:} Grid for counties of Sweden. Image reference \href{https://user-images.githubusercontent.com/6022947/31674182-71150350-b361-11e7-9873-ffe319344c90.jpg}{here}. Thanks to \href{https://github.com/richardohrvall}{richardohrvall}.
}

\itemize{
\item \strong{uk_regions1:} Grid for regions of the UK (aka EU standard NUTS 1 areas). Image reference \href{https://user-images.githubusercontent.com/1275592/32032394-d394b24a-b9bb-11e7-91e8-fe4da69ba005.jpg}{here}. Thanks to \href{https://github.com/paulb20}{paulb20}.
}

\itemize{
\item \strong{us_state_contiguous_grid1:} Grid layout for the contiguous US states (including DC). Image reference \href{https://images7.alphacoders.com/687/687665.jpg}{here}. Thanks to \href{https://github.com/andrewsr}{andrewsr}.
}

\itemize{
\item \strong{sk_province_grid1:} Grid layout for South Korean sis and dos (metropolitan/special/autonomous cities and provinces). Image reference \href{http://tong.visitkorea.or.kr/enu/images/koreastay/korea_map.jpg}{here}. Thanks to \href{https://github.com/heon131}{heon131}.
}

\itemize{
\item \strong{ch_aargau_districts_grid1:} Grid layout for Districts of the Canton of Aargau, Switzerland. Image reference \href{https://user-images.githubusercontent.com/31518032/32281394-fcd761d0-bf1e-11e7-9dcc-ae52307eeafd.png}{here}. Thanks to \href{https://github.com/zumbov2}{zumbov2}.
}

\itemize{
\item \strong{jo_gov_grid1:} Grid layout for Governorates of Jordan. Image reference \href{https://camo.githubusercontent.com/a229e8b82407259265378f39e4b5f09dceace668/687474703a2f2f7777772e3234706f696e74302e636f6d2f7070742d73686f702f6d656469612f636174616c6f672f70726f647563742f63616368652f312f696d6167652f61313437653035336533386461386432366462626233626232346232313238302f6a2f6f2f6a6f7264616e2d676f7665726e6f726174652d6361706974616c732d6d61702d706f776572706f696e742d70726573656e746174696f6e2e6a7067}{here}. Thanks to \href{https://github.com/ghosthedirewolf}{ghosthedirewolf}.
}

\itemize{
\item \strong{spain_ccaa_grid1:} Grid layout for Spanish 'Comunidades Autónomas'. Image reference \href{https://camo.githubusercontent.com/5be0fc76e8dab5269761384ce2b77e7e79bba0db/68747470733a2f2f7777772e6d737373692e676f622e65732f6f7267616e697a6163696f6e2f636361612f696d672f737061696e2e6a7067}{here}. Thanks to \href{https://github.com/JoseAntonioOrtega}{JoseAntonioOrtega}.
}

\itemize{
\item \strong{spain_prov_grid2:} Grid layout for Provinces of Spain. Image reference \href{https://user-images.githubusercontent.com/1275592/30189589-cbfbf2de-93e9-11e7-9598-bb4c4e09e0c1.png}{here}. Thanks to \href{https://github.com/JoseAntonioOrtega}{JoseAntonioOrtega}.
}

\itemize{
\item \strong{world_countries_grid1:} Grid layout for countries of the world, with a few exclusions. See . Image reference \href{https://user-images.githubusercontent.com/1275592/30189589-cbfbf2de-93e9-11e7-9598-bb4c4e09e0c1.png}{here}. Thanks to \href{https://github.com/JoseAntonioOrtega}{JoseAntonioOrtega}.
}

\itemize{
\item \strong{br_states_grid2:} Grid for the 27 states of Brazil. Image reference \href{http://www.brazil-help.com/mapa_do_brasil.gif}{here}. Thanks to \href{https://github.com/hafen}{hafen}.
}

\itemize{
\item \strong{china_city_grid1:} Grid layout of cities in China. Image reference \url{here}. Thanks to \href{https://github.com/CharleneDeng1}{CharleneDeng1}.
}

\itemize{
\item \strong{kr_seoul_district_grid1:} Grid layout of Seoul's 25 districts. Image reference \href{https://user-images.githubusercontent.com/1275592/38068875-494e49a8-32c8-11e8-8ace-b91982fe486f.png}{here}. Thanks to \href{https://github.com/yonghah}{yonghah}.
}

\itemize{
\item \strong{nz_regions_grid1:} Grid layout for regions of New Zealand. Image reference \href{https://camo.githubusercontent.com/3db8d49d196be1e70ed0ebd4dcedf6fefc8fcfed/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f322f32312f4e5a5f526567696f6e616c5f436f756e63696c735f616e645f5465727269746f7269616c5f417574686f7269746965735f323031372e7376672f3130303070782d4e5a5f526567696f6e616c5f436f756e63696c735f616e645f5465727269746f7269616c5f417574686f7269746965735f323031372e7376672e706e67}{here}. Thanks to \href{https://github.com/pierreroudier}{pierreroudier}.
}

\itemize{
\item \strong{sl_regions_grid1:} Grid layout of Slovenian regions. Image reference \href{https://user-images.githubusercontent.com/1275592/38069181-df2cbcb0-32c9-11e8-8a5b-0528f3e9e81c.png}{here}. Thanks to \href{https://github.com/SR1986}{SR1986}.
}

\itemize{
\item \strong{us_census_div_grid1:} Grid layout of US Census divisions. Image reference \href{https://camo.githubusercontent.com/154dc45aaff776642ac7a5ce257106f70c97d441/68747470733a2f2f7777772e6569612e676f762f636f6e73756d7074696f6e2f636f6d6d65726369616c2f696d616765732f63656e646976636f2e676966}{here}. Thanks to \href{https://github.com/mkiang}{mkiang}.
}

\itemize{
\item \strong{ar_tucuman_province_grid1:} Grid layout for Argentina Tucumán Province political divisions (departments) Image reference \href{https://camo.githubusercontent.com/03faf8f653fffe62e0beb2ad9d36cbcbce094c10/687474703a2f2f7777772e636f6d6f6c6c65676172612e636f6d2f696d616765732f617267656e74696e612f646570617274616d656e746f732d64652d6c612d70726f76696e6369612d64652d747563756d616e2e706e67}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{us_nh_counties_grid1:} Grid layout for the 10 counties in New Hampshire. Image reference \href{https://en.wikipedia.org/wiki/List_of_counties_in_New_Hampshire#/media/File:New_Hampshire_Locator_Map.PNG}{here}. Thanks to \href{https://github.com/soungl}{soungl}.
}

\itemize{
\item \strong{china_prov_grid2:} Grid layout for Provinces of China. Image reference \href{https://user-images.githubusercontent.com/36669758/37746305-f5c138f4-2d4f-11e8-8a48-2a21513ea828.png}{here}. Thanks to \href{https://github.com/jw2531}{jw2531}.
}

\itemize{
\item \strong{pl_voivodeships_grid1:} Grid layout for Polish voivodeships (provinces) Image reference \href{https://user-images.githubusercontent.com/5550348/37672220-69c03caa-2c65-11e8-857e-2955f25f55bf.png}{here}. Thanks to \href{https://github.com/erzk}{erzk}.
}

\itemize{
\item \strong{us_ia_counties_grid1:} Grid layout for counties in Iowa Image reference \href{https://camo.githubusercontent.com/9a5e5d3c974c85a6d3ee87b1d2858940476e74c0/68747470733a2f2f7777772e6469676974616c2d746f706f2d6d6170732e636f6d2f636f756e74792d6d61702f696f77612d636f756e74792d6d61702e676966}{here}. Thanks to \href{https://github.com/jrennyb}{jrennyb}.
}

\itemize{
\item \strong{us_id_counties_grid1:} Grid layout for counties in Idaho Image reference \href{https://camo.githubusercontent.com/dbe964d952ecd106d2a9f677950852f957be1281/68747470733a2f2f67656f6c6f67792e636f6d2f636f756e74792d6d61702f696461686f2d636f756e74792d6d61702e676966}{here}. Thanks to \href{https://github.com/hathawayj}{hathawayj}.
}

\itemize{
\item \strong{ar_cordoba_dep_grid1:} Grid layout for departments of Cordoba province in Argentina. Image reference \href{https://user-images.githubusercontent.com/12114624/38060507-45d03c66-32a8-11e8-81c0-3ef7675dbf4e.png}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{us_fl_counties_grid1:} Grid for Florida counties. Image reference \href{https://user-images.githubusercontent.com/1275592/38098066-df3a7690-332b-11e8-9e4f-9123379e412e.jpg}{here}. Thanks to \href{https://github.com/ejr248}{ejr248}.
}

\itemize{
\item \strong{ar_buenosaires_communes_grid1:} Grid for communes of Buenos Aires, Argentina. Image reference \href{https://camo.githubusercontent.com/1397b556f643333f5c0afbf074e12b316ea8f063/687474703a2f2f7777772e6e6f746963696172696f7375722e636f6d2e61722f77702d636f6e74656e742f75706c6f6164732f323031322f30322f434f4d4e41532d4d4150412e676966}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{nz_regions_grid2:} Grid layout for regions of New Zealand. Image reference \href{https://camo.githubusercontent.com/3db8d49d196be1e70ed0ebd4dcedf6fefc8fcfed/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f322f32312f4e5a5f526567696f6e616c5f436f756e63696c735f616e645f5465727269746f7269616c5f417574686f7269746965735f323031372e7376672f3130303070782d4e5a5f526567696f6e616c5f436f756e63696c735f616e645f5465727269746f7269616c5f417574686f7269746965735f323031372e7376672e706e67}{here}. Thanks to \href{https://github.com/pierreroudier}{pierreroudier}.
}

\itemize{
\item \strong{oecd_grid1:} Grid layout for OECD member countries. Image reference \href{https://camo.githubusercontent.com/9afd6ca4c97374e8d5b69c73325937e6de57ade3/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f342f34632f4f4543445f6d656d6265725f7374617465735f6d61702e7376672f3130303070782d4f4543445f6d656d6265725f7374617465735f6d61702e7376672e706e67}{here}. Thanks to \href{https://github.com/arcruz0}{arcruz0}.
}

\itemize{
\item \strong{ec_prov_grid1:} Grid layout for provinces of Ecuador Image reference \href{https://camo.githubusercontent.com/7eabd24db749e5335f13c911df65f15516c10bef/68747470733a2f2f65637561646f726e6f7469636961732e636f6d2f77702d636f6e74656e742f75706c6f6164732f323031372f30342f70726f76696e636961732d65637561646f722e6a7067}{here}. Thanks to \href{https://github.com/Ricardo95RM}{Ricardo95RM}.
}

\itemize{
\item \strong{nl_prov_grid1:} Grid layout for provinces of Netherlands Image reference \href{https://camo.githubusercontent.com/a094b2603eff7e4c45b9fbdbbfd7b5637473354c/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f632f63632f4d61705f70726f76696e6365735f4e65746865726c616e64732d656e2e737667}{here}. Thanks to \href{https://github.com/ruditurksema}{ruditurksema}.
}

\itemize{
\item \strong{ca_prov_grid1:} Grid layout for provinces of Canada Image reference \href{https://camo.githubusercontent.com/5a0ad66e98796b817e506fe3127d48d7277a74a8/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f312f31342f506f6c69746963616c5f6d61705f6f665f43616e6164612e706e672f38303070782d506f6c69746963616c5f6d61705f6f665f43616e6164612e706e67}{here}. Thanks to \href{https://github.com/michael-chong}{michael-chong}.
}

\itemize{
\item \strong{us_nc_counties_grid1:} Grid layout for Counties of North Carolina, United States Image reference \href{https://camo.githubusercontent.com/50bf9a4f4861fa11dba4ea66f8f7712eb1b59593/68747470733a2f2f7765622e6c69622e756e632e6564752f6e632d6d6170732f696d616765732f686f74636f756e74796d61702e676966}{here}. Thanks to \href{https://github.com/mtdukes}{mtdukes}.
}

\itemize{
\item \strong{mx_ciudad_prov_grid1:} Grid layout for Districts of Mexico City, Mexico Image reference \href{https://camo.githubusercontent.com/ef6379d66d92e1ea289fcd8522853ec4348499d7/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f362f36652f4d582d44462d4469766973692543332542336e5f706f6c254333254144746963612e7376672f38303070782d4d582d44462d4469766973692543332542336e5f706f6c254333254144746963612e7376672e706e67}{here}. Thanks to \href{https://github.com/Ivangea}{Ivangea}.
}

\itemize{
\item \strong{bg_prov_grid1:} Grid layout for provinces of Bulgaria Image reference \href{https://camo.githubusercontent.com/46872c5c1703af4739df87eca5075ae91fe83833/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f382f38362f526567696f6e735f6f665f42756c67617269615f4d61702e706e67}{here}. Thanks to \href{https://github.com/savinastoitsova}{savinastoitsova}.
}

\itemize{
\item \strong{us_hhs_regions_grid1:} This grid approximates the U.S. Health and Human Services Region map Image reference \href{https://camo.githubusercontent.com/dd6f1a0c7749dd111d71be74f344d573eaad52f5/68747470733a2f2f7777772e6868732e676f762f73697465732f64656661756c742f66696c65732f6965612f696d616765732f726567696f6e736d61702e6a7067}{here}. Thanks to \href{https://github.com/akitepowell}{akitepowell}.
}

\itemize{
\item \strong{tw_counties_grid1:} Grid layout for Counties of Taiwan Image reference \href{https://camo.githubusercontent.com/b51d87c85c060ff96335274808600f7984f19e56/68747470733a2f2f692e696d6775722e636f6d2f744a5a6a4f4f422e706e67}{here}. Thanks to \href{https://github.com/csh484912274}{csh484912274}.
}

\itemize{
\item \strong{tw_counties_grid2:} Grid layout for Counties of Taiwan including Lienchiang County Image reference \href{https://camo.githubusercontent.com/b26bf4a5ff70294daaa6db508e06a82f2afd902f/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f652f65372f54616977616e5f524f435f706f6c69746963616c5f6469766973696f6e735f6c6162656c65642e7376672f3132303070782d54616977616e5f524f435f706f6c69746963616c5f6469766973696f6e735f6c6162656c65642e7376672e706e67}{here}. Thanks to \href{https://github.com/csh484912274}{csh484912274}.
}

\itemize{
\item \strong{af_prov_grid1:} Grid layout for Provinces of Afghanistan Image reference \href{https://camo.githubusercontent.com/86d2b1807e37f2d3009da212a870b58f9b091d67/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f342f34342f41666768616e697374616e5f70726f76696e6365735f6e756d62657265642e706e67}{here}. Thanks to \href{https://github.com/jrennyb}{jrennyb}.
}

\itemize{
\item \strong{us_mi_counties_grid1:} Grid layout for Counties of Michigan, United States Image reference \href{https://camo.githubusercontent.com/834a8eaa5d2bc37ec003d06d0cb5eee53954320d/687474703a2f2f7777772d706572736f6e616c2e756d6963682e6564752f7e62626f776d616e2f62697264732f4352414d5f696d6167656e756d732e676966}{here}. Thanks to \href{https://github.com/jrennyb}{jrennyb}.
}

\itemize{
\item \strong{pe_prov_grid1:} Grid layout for Provinces of Peru Image reference \href{https://camo.githubusercontent.com/6c8d13fd43bf90b6bc7dbc3bff6c5a2abcda2e52/68747470733a2f2f696d6167656e6573746f74616c65732e636f6d2f77702d636f6e74656e742f75706c6f6164732f323031382f30342f4d6170612d706f6c697469636f2d64656c2d506572752d322d373638783933362e706e67}{here}. Thanks to \href{https://github.com/jmcastagnetto}{jmcastagnetto}.
}

\itemize{
\item \strong{sa_prov_grid2:} Grid layout for Provinces of South Africa Image reference \href{https://camo.githubusercontent.com/14e357c36b6b72c0ef60e700d832b57aac87bddd/68747470733a2f2f7777772e736f7574686166726963612e746f2f70726f76696e6365732f5253415f62795f70726f76696e6365732e6a7067}{here}. Thanks to \href{https://github.com/kamermanpr}{kamermanpr}.
}

\itemize{
\item \strong{mx_state_grid3:} Grid layout for States of Mexico Image reference \href{https://camo.githubusercontent.com/3b62258772065fe17c3a7a6a99b889251fa53d4a/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f632f63322f506f6c69746963616c5f6469766973696f6e735f6f665f4d657869636f2d656e2e7376672f3132303070782d506f6c69746963616c5f6469766973696f6e735f6f665f4d657869636f2d656e2e7376672e706e67}{here}. Thanks to \href{https://github.com/ikashnitsky}{ikashnitsky}.
}

\itemize{
\item \strong{cn_bj_districts_grid1:} Grids for Administrative Districts of Beijing, China Image reference \href{https://user-images.githubusercontent.com/1275592/54295842-0ccea100-4571-11e9-8d3f-c60f4f7f2c0a.png}{here}. Thanks to \href{https://github.com/shiedelweiss}{shiedelweiss}.
}

\itemize{
\item \strong{us_va_counties_grid1:} Grids for Counties of Virginia, United States Image reference \href{https://camo.githubusercontent.com/ad5f91421db6241dae5a292d841c13475b39452e/68747470733a2f2f67656f6c6f67792e636f6d2f73746174652d6d61702f6d6170732f76697267696e69612d636f756e74792d6d61702e676966}{here}. Thanks to \href{https://github.com/joshyazman}{joshyazman}.
}

\itemize{
\item \strong{us_mo_counties_grid1:} Grids for Counties of Missouri, United States Image reference \href{https://camo.githubusercontent.com/d58ebdce931d0d42085d3b4d76f9e809bd147f84/68747470733a2f2f6f736564612e6d6973736f7572692e6564752f636f756e7479706167652f4d4f5f636f756e746965732e706e67}{here}. Thanks to \href{https://github.com/Yanqi-Xu}{Yanqi-Xu}.
}

\itemize{
\item \strong{cl_santiago_prov_grid1:} Communes of Santiago Province, Chile Image reference \href{https://camo.githubusercontent.com/64e592a3c79ce5b03024cccfe43a9fca64ebffe1/68747470733a2f2f692e696d6775722e636f6d2f457361444675762e706e67}{here}. Thanks to \href{https://github.com/robsalasco}{robsalasco}.
}

\itemize{
\item \strong{us_tx_capcog_counties_grid1:} This is a grid of a 10 county planning region around Austin, Texas, United States Image reference \href{https://camo.githubusercontent.com/029adb9992cb52f8b655935a8ea67f3b2ba2778f/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f322f32332f434150434f472e706e67}{here}. Thanks to \href{https://github.com/mth444}{mth444}.
}

\itemize{
\item \strong{sg_planning_area_grid1:} Grids for Planning Areas of Singapore Image reference \href{https://user-images.githubusercontent.com/1275592/54297744-a9467280-4574-11e9-99f7-bdb62bbb34b4.png}{here}. Thanks to \href{https://github.com/Elenafuyi}{Elenafuyi}.
}

\itemize{
\item \strong{in_state_ut_grid1:} Grid of Indian States and Union Territories Image reference \href{https://user-images.githubusercontent.com/1275592/54297966-1f4ad980-4575-11e9-8e8c-5006c91fccd1.png}{here}. Thanks to \href{https://github.com/seanangio}{seanangio}.
}

\itemize{
\item \strong{cn_fujian_prov_grid1:} Grid of counties of Fujian Province, China Image reference \href{https://user-images.githubusercontent.com/1275592/54298318-c2035800-4575-11e9-8d68-868cc76315df.png}{here}. Thanks to \href{https://github.com/nannanchen333}{nannanchen333}.
}

\itemize{
\item \strong{ca_quebec_electoral_districts_grid1:} Grid of Electoral Districts of Québec, Canada Image reference \href{https://camo.githubusercontent.com/d1f75185f1a675f1a358297ad0dbdae1b0d7b131/68747470733a2f2f7777772e656c656374696f6e737175656265632e71632e63612f3730616e732f6173736574732f696d672f323031312d342e6a7067}{here}. Thanks to \href{https://github.com/jhroy}{jhroy}.
}

\itemize{
\item \strong{nl_prov_grid2:} Grid with the provinces of The Netherlands with codes that are used by the statistical institute of NL Image reference \href{https://camo.githubusercontent.com/2b6204b5f9856fadeb32a6042fe1fd48cbb4b06f/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f332f33302f4e656465726c616e64736550726f76696e636965734c617267652e706e67}{here}. Thanks to \href{https://github.com/edwindj}{edwindj}.
}

\itemize{
\item \strong{cn_bj_districts_grid2:} Grid with districts of Beijing, China Image reference \href{https://camo.githubusercontent.com/c32d419c0b3d10fa007e68a66bf368752a45f8a9/68747470733a2f2f677373302e62647374617469632e636f6d2f2d346f33645361675f7849346b68476b706f574b314846366868792f6261696b652f63302533446261696b65313530253243352532433525324331353025324335302f7369676e3d34396436313237373464313062393132616263636665616361323934393736362f643034336164346264313133373366303036633533613533616530663462666266616564303430372e6a7067}{here}. Thanks to \href{https://github.com/zouhx11}{zouhx11}.
}

\itemize{
\item \strong{ar_santiago_del_estero_prov_grid1:} Grid with districts of Santiago del Estero Province, Argentina Image reference \href{https://camo.githubusercontent.com/82e68c0a0f0b243c8a9d21133227f74e05c93c6b/68747470733a2f2f74342e6b6e332e6e65742f746172696e67612f302f382f312f362f362f312f4d6565656b69692f3631382e6a7067}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_formosa_prov_grid1:} Grid with districts of Formosa Province, Argentina Image reference \href{https://camo.githubusercontent.com/481f581ddcd2ae7a09cd47fb4779dc904a0d6b75/687474703a2f2f7777772e74656d7075736167726f2e636f6d2e61722f696d672f666f726d6f73612f6469766973696f6e706f6c69746963612e6a7067}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_chaco_prov_grid1:} Grid with districts of Chaco Province, Argentina Image reference \href{https://camo.githubusercontent.com/35120f89439dbdd5c147d3925d948ae6a955b0c8/68747470733a2f2f657370616e6f6c2e6d6170736f66776f726c642e636f6d2f77702d636f6e74656e742f75706c6f6164732f323031342f30362f6d6170612d64656c2d636861636f2e6a7067}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_catamarca_prov_grid1:} Grid with districts of Catamarca Province, Argentina Image reference \href{https://camo.githubusercontent.com/8a25c2986d42003208644618e0e3d3c1950fcd92/687474703a2f2f7777772e616c6c2d617267656e74696e612e636f6d2e61722f77702d636f6e74656e742f75706c6f6164732f6d6170612d706f6c697469636f2d636174616d617263612e6a7067}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_jujuy_prov_grid1:} Grid with districts of Jujuy Province, Argentina Image reference \href{https://camo.githubusercontent.com/df150ce3e47d85c63efc43a5614f5f6366970669/687474703a2f2f7777772e64697065632e6a756a75792e676f762e61722f6d617061735f6a756a75795f63656e736f2f6a756a7579646570617274616d656e746f732e676966}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_neuquen_prov_grid1:} Grid with districts of Neuquen Province, Argentina Image reference \href{https://camo.githubusercontent.com/5cda1ad7e1b4a39c6932a9c4cbfe7f2dc7cd9cd6/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f662f66322f4d6170615f64655f64656e73696461645f64655f706f626c6163696f6e5f323031305f70726f76696e6369615f64656c5f6e65757175656e2e7376672f32353070782d4d6170615f64655f64656e73696461645f64655f706f626c6163696f6e5f323031305f70726f76696e6369615f64656c5f6e65757175656e2e7376672e706e67}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_san_luis_prov_grid1:} Grid with districts of San Luis Province, Argentina Image reference \href{https://camo.githubusercontent.com/a1c246c3e3b95e7186355a5ac473bb68ac2818b4/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f612f61342f4469766973692543332542336e5f706f6c254333254144746963615f64655f6c615f50726f76696e6369615f64655f53616e5f4c7569735f2d417267656e74696e612d2e6a7067}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_san_juan_prov_grid1:} Grid with districts of San Juan Province, Argentina Image reference \href{https://camo.githubusercontent.com/a29a04aa5dbd26fd425abe01f07a60bb6d563013/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f302f30652f4d6170615f64655f6c615f50726f76696e6369615f64655f53616e5f4a75616e5f253238417267656e74696e612532395f636f6e5f6469766973692543332542336e5f706f6c254333254144746963612e6a70672f35303070782d4d6170615f64655f6c615f50726f76696e6369615f64655f53616e5f4a75616e5f253238417267656e74696e612532395f636f6e5f6469766973692543332542336e5f706f6c254333254144746963612e6a7067}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_santa_fe_prov_grid1:} Grid with districts of Santa Fe Province, Argentina Image reference \href{https://camo.githubusercontent.com/e3812264494cb18796173767991749601b0256d6/68747470733a2f2f7777772e66616d696c797365617263682e6f72672f77696b692f65732f696d675f617574682e7068702f7468756d622f372f37642f53616e74615f46655f50726f76696e6369615f4d6170612e706e672f33373070782d53616e74615f46655f50726f76696e6369615f4d6170612e706e67}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_la_rioja_prov_grid1:} Grid with districts of La Rioja Province, Argentina Image reference \href{https://camo.githubusercontent.com/03c888b6d01726c6832975a1535a046d91cd26ab/68747470733a2f2f7777772e66616d696c797365617263682e6f72672f77696b692f65732f696d675f617574682e7068702f7468756d622f642f64612f4c615f52696f6a615f50726f76696e6369615f4d6170612e706e672f36323070782d4c615f52696f6a615f50726f76696e6369615f4d6170612e706e67}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_mendoza_prov_grid1:} Grid with districts of Mendoza Province, Argentina Image reference \href{https://camo.githubusercontent.com/83ccf5cfdde385d5aa502812b61efafdcc769ecb/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f332f33342f4d6170615f64655f6c615f6469766973692543332542336e5f706f6c254333254144746963615f64655f6c615f706369612e5f64655f4d656e6f7a615f2d417267656e74696e612d2e6a70672f38303070782d4d6170615f64655f6c615f6469766973692543332542336e5f706f6c254333254144746963615f64655f6c615f706369612e5f64655f4d656e6f7a615f2d417267656e74696e612d2e6a7067}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_salta_prov_grid1:} Grid with districts of Salta Province, Argentina Image reference \href{https://camo.githubusercontent.com/9232d808104ddd3ab3983f9a670eb507a4d95e51/687474703a2f2f312e62702e626c6f6773706f742e636f6d2f5f564a6b71374771783743512f52303349467549766e30492f41414141414141414141632f75336655614c744e6a53672f733430302f646570617274616d656e746f732b64652b53616c74612e6a7067}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_rio_negro_prov_grid1:} Grid with districts of Rio Negro Province, Argentina Image reference \href{https://camo.githubusercontent.com/123414ec4933d38f3b991cc8d669138594d68b17/68747470733a2f2f63646e2e7468696e676c696e6b2e6d652f6170692f696d6167652f3536353530363334383233343537313737362f313234302f31302f7363616c65746f7769647468}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{uy_departamentos_grid1:} Grid with Departamentos of Uruguay Image reference \href{https://camo.githubusercontent.com/a56f891616c613ba8c120434f3d0786e005f1df8/68747470733a2f2f6d6170616d756e64692e6f6e6c696e652f77702d636f6e74656e742f75706c6f6164732f323031382f31312f6d6170612d706f6c697469636f2d64652d757275677561792e6a7067}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{ar_buenos_aires_prov_electoral_dist_grid1:} Grid with Electoral Districts of Buenos Aires Province, Argentina Image reference \href{https://camo.githubusercontent.com/3803a78b6fc2ca41ffd7bdf3a7a503af1cb48ba7/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f7468756d622f622f62302f53656363696f6e65735f656c6563746f72616c65735f64655f6c615f70726f76696e6369615f64655f4275656e6f735f41697265732e7376672f38303070782d53656363696f6e65735f656c6563746f72616c65735f64655f6c615f70726f76696e6369615f64655f4275656e6f735f41697265732e7376672e706e67}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{europe_countries_grid1:} Grid layout for all European countries except Vatican City, Monaco, San Marino and Liechtenstein Image reference \href{https://user-images.githubusercontent.com/38204689/38485856-23e709b6-3bd3-11e8-8bc7-ae5cd892419c.png}{here}. Thanks to \href{https://github.com/THargreaves}{THargreaves}.
}

\itemize{
\item \strong{argentina_grid2:} Grid layout for Argentina without Islas Malvinas and Antártida Argentina Image reference \href{https://camo.githubusercontent.com/f14f25ac3e42051994f5fd4c53e2718c2b2290aa/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f392f39652f4d61705f6f665f417267656e74696e615f776974685f70726f76696e6365735f6e616d65735f656e2e706e67}{here}. Thanks to \href{https://github.com/TuQmano}{TuQmano}.
}

\itemize{
\item \strong{us_state_grid5:} Grid layout for United States with AK and HI flush with CA Image reference \href{https://camo.githubusercontent.com/b446d99f631d067562f50a57449eddb0b89c1acf/68747470733a2f2f75706c6f61642e77696b696d656469612e6f72672f77696b6970656469612f636f6d6d6f6e732f612f61352f4d61705f6f665f5553415f776974685f73746174655f6e616d65732e737667}{here}. Thanks to \href{https://github.com/christophercannon}{christophercannon}.
}

\itemize{
\item \strong{jp_prefs_grid2:} Grid layout for Prefectures of Japan Image reference \href{https://user-images.githubusercontent.com/19657164/46902315-4c2f8d80-cefe-11e8-90bb-661771ce2836.JPG}{here}. Thanks to \href{https://github.com/Ryo-N7}{Ryo-N7}.
}
}
\keyword{data}
