% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{india_grid1}
\alias{india_grid1}
\title{india_grid1}
\usage{
india_grid1
}
\description{
A data frame containing a grid layout for India states, thanks to \url{https://github.com/meysubb}.
}
\keyword{data}
