% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{india_pop}
\alias{india_pop}
\title{india_pop}
\usage{
india_pop
}
\description{
2011 population data for India, broken down by urban and rural. Source: \url{https://en.wikipedia.org/wiki/List_of_states_and_union_territories_of_India_by_population}.
}
\keyword{data}
