% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{london_afford}
\alias{london_afford}
\title{london_afford}
\usage{
london_afford
}
\description{
Total affordable housing completions by financial year in each London borough since 2015/16. Source: \url{https://data.london.gov.uk/dataset/dclg-affordable-housing-supply-borough}
}
\keyword{data}
