% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{london_boroughs_grid}
\alias{london_boroughs_grid}
\title{london_boroughs_grid}
\usage{
london_boroughs_grid
}
\description{
A data frame containing a grid layout for the boroughs of London, thanks to \url{https://github.com/eldenvo}. Note that the column \code{code_ons} contains the \href{http://geoportal.statistics.gov.uk/datasets/464be6191a434a91a5fa2f52c7433333_0?uiTab=table}{codes} used by UK Office for National Statistics.
}
\keyword{data}
