% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{nhs_scot_dental}
\alias{nhs_scot_dental}
\title{nhs_scot_dental}
\usage{
nhs_scot_dental
}
\description{
Child dental health data in Scotland. Source: \url{http://statistics.gov.scot/data/child-dental-health}
}
\keyword{data}
