% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{nhs_scot_grid}
\alias{nhs_scot_grid}
\title{nhs_scot_grid}
\usage{
nhs_scot_grid
}
\description{
A data frame containing a grid layout for a grid of NHS Scotland Health Boards, thanks to \url{https://gist.github.com/josephjosephadams}. Note that the column \code{code} contains the \href{http://geoportal.statistics.gov.uk/datasets/61dca36aa33e4aee99bbfa784fa9eb25_0?uiTab=table}{codes} used by UK Office for National Statistics.
}
\keyword{data}
