% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_geo.R
\name{plot.facet_geo}
\alias{plot.facet_geo}
\title{Plot geofaceted ggplot2 object}
\usage{
\method{plot}{facet_geo}(x, ...)
}
\arguments{
\item{x}{plot object}

\item{...}{ignored}
}
\description{
Plot geofaceted ggplot2 object
}
