% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_geo.R
\name{+.gg}
\alias{+.gg}
\title{Add method for gg / facet_geo}
\usage{
\method{+}{gg}(e1, e2)
}
\arguments{
\item{e1}{a object with class gg}

\item{e2}{if object is of class 'facet_geo', then 'facet_geo' will be appended to the class of e1}
}
\description{
Add method for gg / facet_geo
}
