% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_geo.R
\name{print.facet_geo}
\alias{print.facet_geo}
\title{Print geofaceted ggplot2 object}
\usage{
\method{print}{facet_geo}(x, ...)
}
\arguments{
\item{x}{plot object}

\item{...}{ignored}
}
\description{
Print geofaceted ggplot2 object
}
