% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{sa_pop_dens}
\alias{sa_pop_dens}
\title{sa_pop_dens}
\usage{
sa_pop_dens
}
\description{
Population density for each province in South Africa for 1996, 2001, and 2011. Source: \url{https://en.wikipedia.org/wiki/List_of_South_African_provinces_by_population_density}
}
\keyword{data}
