% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{sa_prov_grid1}
\alias{sa_prov_grid1}
\title{sa_prov_grid1}
\usage{
sa_prov_grid1
}
\description{
A data frame containing a grid layout for the provinces of South Africa, thanks to \url{https://github.com/jonmcalder}.
}
\keyword{data}
