% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{state_unemp}
\alias{state_unemp}
\title{state_unemp}
\usage{
state_unemp
}
\description{
Seasonally-adjusted December unemployment rate for each state (including DC) from 2000 to 2017. Obtained from bls.gov.
}
\keyword{data}
