% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_geo.R
\name{submit_grid}
\alias{submit_grid}
\title{Print geofaceted ggplot}
\usage{
submit_grid(x, name = NULL, desc = NULL)
}
\arguments{
\item{x}{a data frame containing a grid}

\item{name}{proposed name of the grid (if not supplied, will be asked for interactively)}

\item{desc}{a description of the grid (if not supplied, will be asked for interactively)}
}
\description{
Print geofaceted ggplot
}
\details{
This opens up a github issue for this package in the web browser with pre-populated content for adding a grid to the package.
}
\examples{
\dontrun{
my_grid <- us_state_grid1
my_grid$col[my_grid$label == "WI"] <- 7
submit_grid(my_grid, name = "us_grid_tweak_wi",
  desc = "Modified us_state_grid1 to move WI over")
}


}
