% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{us_state_grid1}
\alias{us_state_grid1}
\title{us_state_grid1}
\usage{
us_state_grid1
}
\description{
A data frame containing a grid layout for US states (including DC)
}
\keyword{data}
