% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_auto.R
\name{attach_spdf}
\alias{attach_spdf}
\title{Attach a SpatialPolygonsDataFrame object to a grid}
\usage{
attach_spdf(x, spdf)
}
\arguments{
\item{x}{object to attach SpatialPolygonsDataFrame object to}

\item{spdf}{a SpatialPolygonsDataFrame object to attach}
}
\description{
Attach a SpatialPolygonsDataFrame object to a grid
}
