% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-geofacet.R
\docType{data}
\name{election}
\alias{election}
\title{election}
\usage{
election
}
\description{
2016 US presidential election results, obtained from \url{http://bit.ly/2016votecount}.
}
\keyword{data}
