% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_geo.R
\name{grid_preview}
\alias{grid_preview}
\title{Plot a preview of a grid}
\usage{
grid_preview(x, label = NULL)
}
\arguments{
\item{x}{a data frame containing a grid}

\item{label}{the column should be used for text labels}
}
\description{
Plot a preview of a grid
}
\examples{
grid_preview(us_state_grid2)
grid_preview(eu_grid1, label = "name")
}
