% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_geo.R
\name{print.facet_geo}
\alias{print.facet_geo}
\title{Print geofaceted ggplot2 object}
\usage{
\method{print}{facet_geo}(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
\item{x}{plot object}

\item{newpage}{draw new (empty) page first?}

\item{vp}{viewport to draw plot in}

\item{...}{other arguments not used by this method}
}
\description{
Print geofaceted ggplot2 object
}
